// src/core/redux/types/profile.enum.ts
export enum EProfileType {
	CREATE_PROFILE = 'PROFILE/CREATE_PROFILE',
	PROFILE_CREATED = 'PROFILE/PROFILE_CREATED',

	EDIT_PROFILE = 'PROFILE/EDIT_PROFILE',
	PROFILE_UPDATED = 'PROFILE/PROFILE_UPDATED',

	FETCH_PROFILE = 'PROFILE/FETCH_PROFILE',
	PROFILE_FETCHED = 'PROFILE/PROFILE_FETCHED',

	FETCH_PROFILES = 'PROFILE/FETCH_PROFILES',
	PROFILES_FETCHED = 'PROFILE/PROFILES_FETCHED',
}